Fixed log validation

+++ Logging.java
@@ -3,7 +3,7 @@
 import java.util.InputMismatchException;
 import java.util.Arrays;
 import java.util.Scanner;
-import java.util.stream.IntStream;
+import java.util.stream.DoubleStream;
 
 /**
  * Enterprise software for ATREE Logging Company.
@@ -68,7 +68,7 @@ public final class Logging {
       scanner.close();
 
       // Check if the log length is a valid length
-      if (IntStream.of(LOG_LENGTHS).anyMatch(num -> num == logLength)) {
+      if (DoubleStream.of(LOG_LENGTHS).anyMatch(num -> num == logLength)) {
         // Calculate the amount of logs that can fit in the truck
         final double logAmount = TRUCK_WEIGHT_LIMIT
             / (WEIGHT_PER_METER * logLength);
